+++ src/components/content/patients/patients-table-helper.tsx
@@ -1,0 +1,27 @@
+import { Argument } from "classnames";
+import { MinRecordItem } from "@/components/core/table/table-helpers";
+import { PatientModel } from "@/fhir/models";
+
+export type PatientNameColumnProps = {
+  patient: PatientModel;
+};
+
+export const PatientNameColumn = ({ patient }: PatientNameColumnProps) => (
+  <div className="ctw-flex ctw-items-center">
+    <div className="ctw-ml-4">
+      <div className="ctw-flex ctw-font-medium">
+        <div className="ctw-max-w-xs">{patient.fullName}</div>
+        {patient.gender && <div className="ctw-uppercase"> ({patient.gender[0]})</div>}
+      </div>
+      <div className="ctw-text-content-lighter">
+        {patient.dob} ({patient.age})
+      </div>
+    </div>
+  </div>
+);
+
+// Set of props that are optional configurations for the table.
+export type TableOptionProps<T extends MinRecordItem> = {
+  getRowClasses?: (row: T) => Argument; // Adds a row hover effect and calls onClick.
+  onRowClick?: (row: T) => void;
+};