+++ src/services/fqs/queries/allergies.ts
@@ -3,6 +3,7 @@ import { gql } from "graphql-request";
 import {
   fragmentCoding,
   fragmentEncounterReference,
+  fragmentOrganization,
   fragmentPatient,
   fragmentPractitioner,
 } from "./fragments";
@@ -19,6 +20,7 @@ export interface AllergyGraphqlResponse {
 
 export const allergyQuery = gql`
   ${fragmentCoding}
+  ${fragmentOrganization}
   ${fragmentPatient}
   ${fragmentEncounterReference}
   ${fragmentPractitioner}